Add user data seeding and password hashing functionality
This commit adds a new file `userData.js` in the `seeds` directory, which contains an array of user data to be seeded into the database. It also includes a function `seedUserData` that deletes existing user data and inserts the new data into the database. Additionally, a new file `helpers.js` in the `utils` directory is added, which includes a function `hashPassword` that uses bcrypt to hash passwords. Lastly, the `main.handlebars` file in the `views/layouts` directory is modified to include a navigation bar with links for login, signup, and logout, based on the user's authentication status.